Convert script to python and add some features
Dropbox installation may be located in a folder other than the default
(`$HOME/.dropbox-dist`) so a config for its path is added. Because of the
need to add this configuration, the format for the config file must be
change. JSON is chosen because data structure can be nested and its
availabilty. A Windows inf file style configuration is also considered but
does not support bare key-value entries.

There is no need for `$XDG_CONFIG_HOME/dropboxes` folder so this one is change
to a config file.